Rework the model
