+++ src/core-enumeration/Enumeration.Extensions/Enumeration/Enumeration.cs
