+++ src/core-enumeration/Enumeration/Enumeration.csproj
@@ -7,11 +7,11 @@
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
 		<PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
 		<PackageLicenseFile>LICENSE</PackageLicenseFile>
-		<Description>PrimeFuncPack Core Enumeration Extensions</Description>
+		<Description>PrimeFuncPack Core Enumeration</Description>
 		<Copyright>Copyright © 2021 Andrei Sergeev, Pavel Moskovoy</Copyright>
 		<RootNamespace>System</RootNamespace>
-		<AssemblyName>EarlyFuncPack.Core.Enumeration.Extensions</AssemblyName>
-		<Version>0.0.1</Version>
+		<AssemblyName>EarlyFuncPack.Core.Enumeration</AssemblyName>
+		<Version>0.0.5</Version>
 	</PropertyGroup>
 
 	<ItemGroup>
@@ -22,8 +22,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <PackageReference Include="EarlyFuncPack.Core.Enumeration" Version="0.0.1" />
-	  <PackageReference Include="PrimeFuncPack.Core.Optional" Version="1.1.0" />
+	  <PackageReference Include="PrimeFuncPack.Primitives" Version="1.1.1" />
 	</ItemGroup>
 
 </Project>
