+++ src/core-enumeration/Enumeration/Enumeration/Enumeration.EnumerationSet.cs
@@ -1,13 +1,14 @@
 #nullable enable
 
 using System.Collections.Generic;
+using System.Linq;
 using System.Runtime.CompilerServices;
 
 namespace System
 {
     partial class Enumeration<TEnumeration>
     {
-        private static class EnumerationSet
+        internal static class EnumerationSet
         {
             public static IReadOnlySet<TEnumeration> Value => Internal.Value;
 
@@ -16,9 +17,9 @@ namespace System
                 public static IReadOnlySet<TEnumeration> Value = Build();
 
                 [MethodImpl(MethodImplOptions.AggressiveInlining)]
-                private static HashSet<TEnumeration> Build()
+                private static IReadOnlySet<TEnumeration> Build()
                     =>
-                    new HashSet<TEnumeration>(EnumerationCollection.Value);
+                    EnumerationSource<TEnumeration>.Value.ToHashSet();
             }
         }
     }
