+++ src/core-enumeration/Enumeration/Enumeration/Enumeration.ValueSet.cs
