+++ src/core-enumeration/Enumeration/Enumeration/Enumeration.Values.cs
