+++ src/core-enumeration/Enumeration/Enumeration/Enumeration.cs
@@ -1,0 +1,8 @@
+#nullable enable
+
+namespace System
+{
+    public static partial class Enumeration<TEnumeration> where TEnumeration : struct
+    {
+    }
+}
