--- src/enumeration/Enumeration/Enumeration/Enumeration.cs
@@ -1,9 +1,0 @@
-#nullable enable
-
-namespace System
-{
-    public static partial class Enumeration<TEnumeration>
-        where TEnumeration : struct, IEquatable<TEnumeration>, IEquatable<string>
-    {
-    }
-}
